scrapers/scrape_courts.py
import requests
from bs4 import BeautifulSoup
import pandas as pd
from datetime import datetime
import time
import os

class CourtScraper:
    def __init__(self):
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        })
    
    def scrape_indian_kanoon(self, query="land acquisition"):
        # Indian Kanoon Scraper Skeleton
        # Note: They have valid/strict rate limits. Respect robots.txt.
        print(f"Searching Indian Kanoon for '{query}'...")
        base_url = "https://indiankanoon.org/search/"
        params = {
            'formInput': query,
            'pagenum': 1
        }
        
        cases = []
        try:
            # response = self.session.get(base_url, params=params)
            # soup = BeautifulSoup(response.content, 'html.parser')
            # results = soup.find_all('div', class_='result_title')
            pass
        except Exception as e:
            print(f"Error accessing Indian Kanoon: {e}")

        # Dummy data for skeleton
        cases.append({
            'court': 'Supreme Court of India',
            'title': 'Union of India vs Land Owners Association',
            'date': '2023-11-15',
            'citation': '2023 SC 1234',
            'url': 'https://indiankanoon.org/doc/123456/',
            'scraped_at': datetime.now().isoformat()
        })
        
        return cases

    def scrape_sc_judgments(self):
        # Supreme Court Judgment Scraper
        url = "https://main.sci.gov.in/judgments"
        # Requires handling dynamic content/JS often, might need Selenium/Playwright
        pass

if __name__ == "__main__":
    scraper = CourtScraper()
    cases = scraper.scrape_indian_kanoon()
    
    output_dir = "data"
    if not os.path.exists(output_dir):
        os.makedirs(output_dir)

    df = pd.DataFrame(cases)
    output_file = os.path.join(output_dir, 'court_judgments.csv')
    df.to_csv(output_file, index=False)
    print(f"Scraped {len(cases)} cases. Saved to {output_file}")
