requirements.txt
requests
beautifulsoup4
pandas
sqlalchemy
selenium
webdriver-manager
pdfplumber
openpyxl
geopy
python-dotenv
supabase
